requirements.txt
# ================================
# Requirements pour Prefect + ETL
# ================================

# Orchestrateur
prefect>=2.14.0

# PostgreSQL
psycopg2-binary
sqlalchemy>=2.0

# DBT (Postgres)
dbt-core>=1.7
dbt-postgres>=1.7
Template

# Manipulation fichiers + parquet
pandas
pyarrow
fastparquet

# JSON
orjson

# Logging structuré
loguru

# Utils
python-dotenv
pathlib
tqdm
tenacity

# Pour COPY en CSV tabulé
pytz

# Pour compatibilité subprocess + routines sys
typing-extensions

# tests/requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
